tests.functional: Don't rely on PATH.

+++ tests/functional/utils.py
@@ -1,12 +1,12 @@
-
+import copy
 import json
+import os
 import pipes
 import shutil
 import subprocess
 import sys
 import tempfile
 import time
-import os
 
 import infinit.beyond
 import infinit.beyond.bottle
@@ -16,15 +16,19 @@ from datetime import timedelta
 
 class TemporaryDirectory:
 
-  def __init__(self):
-    self.__dir = None
+  def __init__(self, path = None):
+    self.__dir = path
+    self.__del = False
 
   def __enter__(self):
-    self.__dir = tempfile.mkdtemp()
+    if self.__dir is None:
+      self.__dir = tempfile.mkdtemp()
+      self.__del = True
     return self
 
   def __exit__(self, *args, **kwargs):
-    shutil.rmtree(self.__dir)
+    if self.__del:
+      shutil.rmtree(self.__dir)
 
   def __str__(self):
     return str(self.__dir)
@@ -43,34 +47,56 @@ def unreachable():
 
 class Infinit(TemporaryDirectory):
 
-  def __init__(self, beyond = None, infinit_root = None):
+  def __init__(self,
+               beyond = None,
+               infinit_root = None,
+               home = None,
+               user = None):
+    super().__init__(home)
     self.__beyond = beyond
     self.__infinit_root = infinit_root or ''
+    self.__user = user
+
+  def __enter__(self):
+    super().__enter__()
+    return self
 
   @property
   def version(self):
     return self.run(['infinit-volume', '--version'])[0]
-  def run(self, args, input = None, return_code = 0, env = {}):
+
+  @property
+  def user(self):
+    return self.__user
+
+  def spawn(self, args, input = None, return_code = 0, env = {}):
     if isinstance(args, str):
       args = args.split(' ')
+    if '/' not in args[0]:
+      args[0] = 'bin/%s' % args[0]
+    build_dir = os.environ.get('BUILD_DIR')
+    if build_dir:
+      args[0] = '%s/%s' % (build_dir, args[0])
     args[0] += os.environ.get('EXE_EXT', '')
-    self.env = {
-      'PATH': os.environ['PATH'],
-      'INFINIT_HOME': self.dir,
+    env_ = {
       'INFINIT_RDV': '',
       'INFINIT_BACKTRACE': '1',
     }
+    if self.dir is not None:
+      env_['INFINIT_HOME'] = self.dir
+    if self.__user is not None:
+      env_['INFINIT_USER'] = self.__user
     if 'WINEDEBUG' in os.environ:
-        self.env['WINEDEBUG'] = os.environ['WINEDEBUG']
+      env_['WINEDEBUG'] = os.environ['WINEDEBUG']
     if 'ELLE_LOG_LEVEL' in os.environ:
-      self.env['ELLE_LOG_LEVEL'] = os.environ['ELLE_LOG_LEVEL']
+      env_['ELLE_LOG_LEVEL'] = os.environ['ELLE_LOG_LEVEL']
     if self.__beyond is not None:
-      self.env['INFINIT_BEYOND'] = self.__beyond.domain
-    self.env.update(env)
+      env_['INFINIT_BEYOND'] = self.__beyond.domain
+    env_.update(env)
     if input is not None:
       args.append('-s')
     pretty = '%s %s' % (
-      ' '.join('%s=%s' % (k, v) for k, v in self.env.items()),
+      ' '.join('%s=%s' % (k, v) for k, v in env_.items()),
       ' '.join(pipes.quote(arg) for arg in args))
     if input is not None:
       if isinstance(input, list):
@@ -83,7 +109,7 @@ class Infinit(TemporaryDirectory):
     print(pretty)
     process = subprocess.Popen(
       args,
-      env = self.env,
+      env = env_,
       stdin =  subprocess.PIPE,
       stdout =  subprocess.PIPE,
       stderr =  subprocess.PIPE,
@@ -92,16 +118,22 @@ class Infinit(TemporaryDirectory):
     if input is not None:
       # FIXME: On OSX, if you spam stdin before the FDStream takes it
       # over, you get a broken pipe.
-      time.sleep(0.5)
-    out, err = process.communicate(input)
+      process.stdin.write(input)
+    process.pretty = pretty
+    return process
+
+  def run(self, args, input = None, return_code = 0, env = {}):
+    process = self.spawn(args, input, return_code, env)
+    out, err = process.communicate()
     process.wait()
     if process.returncode != return_code:
       reason = err.decode('utf-8')
       print(reason, file = sys.stderr)
       #if process.returncode not in [0, 1]:
       #  unreachable()
-      raise Exception('command failed with code %s: %s (reason: %s)' % \
-                      (process.returncode, pretty, reason))
+      raise Exception(
+        'command failed with code %s: %s (reason: %s)' % \
+        (process.returncode, process.pretty, reason))
     out = out.decode('utf-8')
     self.last_out = out
     self.last_err = err.decode('utf-8')
@@ -300,9 +332,9 @@ class User():
     self.infinit = infinit
 
   def run(self, cli, **kargs):
-    print('run as %s:\t' % self.name, cli)
-    return self.infinit.run(cli.split(' '),
-                            env = { 'INFINIT_USER': self.name }, **kargs)
+    return self.infinit.run(
+      cli.split(' '),
+      env = { 'INFINIT_USER': self.name }, **kargs)
 
   def run_split(self, args, **kargs):
     return self.infinit.run(args, env = { 'INFINIT_USER': self.name }, **kargs)
